backend/requirements.txt
coverage==5.1
Django==2.2.14
djangorestframework==3.11.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
